+++ examples/tailwindcss/pages/_app.tsx
@@ -1,6 +1,6 @@
 import '../styles/globals.css';
 import type { AppProps } from 'next/app';
-import { defaultTheme, ChainProvider } from '@cosmos-kit/react';
+import { ChainProvider } from '@cosmos-kit/react';
 import { wallets as keplrWallets } from '@cosmos-kit/keplr';
 import { wallets as cosmostationWallets } from '@cosmos-kit/cosmostation';
 import { wallets as leapWallets } from '@cosmos-kit/leap';
@@ -10,10 +10,11 @@ import { ThemeProvider } from '../contexts/theme';
 
 import { SignerOptions } from '@cosmos-kit/core';
 import { chains, assets } from 'chain-registry';
+import '@interchain-ui/react/styles';
 
 function CreateCosmosApp({ Component, pageProps }: AppProps) {
   const signerOptions: SignerOptions = {
-    // signingStargate: (_chain: Chain) => {
+    // signingStargate: () => {
     //   return getSigningCosmosClientOptions();
     // }
   };
@@ -35,7 +36,6 @@ function CreateCosmosApp({ Component, pageProps }: AppProps) {
           },
         },
       }}
-      wrappedWithChakra={true}
       signerOptions={signerOptions}
       walletModal={TailwindModal}
     >