+++ packages/create-cosmos-app/types/git-cca-template.d.ts
@@ -1,1 +1,0 @@
-export declare const createGitApp: (repo: string) => (argv: any) => Promise<any>;