+++ packages/create-cosmos-app/types/git-question-template.d.ts
@@ -1,1 +1,0 @@
-export declare const createQuestionTemplate: (repo: string) => (argv: any) => Promise<any>;