+++ packages/create-osmosis-app/src/cli.ts
@@ -1,7 +1,0 @@
-import { createGitApp } from "create-cosmos-app";
-const createCosmosApp = createGitApp('https://github.com/cosmology-tech/create-cosmos-app.git');
-export const cli = async (argv) => {
-  argv.example = true;
-  argv.template = 'osmosis';
-  await createCosmosApp(argv);
-};