+++ packages/create-osmosis-app/types/git-cca-template.d.ts
@@ -1,1 +1,0 @@
-export declare const createApp: (repo: string) => (argv: any) => Promise<any>;