+++ examples/vote-proposal/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.2](https://github.com/cosmology-tech/create-cosmos-app/compare/@cosmology/vote-proposal@0.2.1...@cosmology/vote-proposal@0.2.2) (2023-07-14)
+
+**Note:** Version bump only for package @cosmology/vote-proposal
+
+
+
+
+
 ## [0.2.1](https://github.com/cosmology-tech/create-cosmos-app/compare/@cosmology/vote-proposal@0.2.0...@cosmology/vote-proposal@0.2.1) (2023-06-28)
 
 **Note:** Version bump only for package @cosmology/vote-proposal